pyproject.toml
[tool.black]
target-version = ['py39']
exclude = '''
/(\.git | \.cache | \.pytest_cache | venv | coverage)/
'''

[tool.ruff]
line-length = 88

[tool.isort]
profile = 'black'
multi_line_output = 3
skip = ['.cache', '.pytest_cache', 'venv', 'coverage', '.git']

[tool.coverage.run]
omit = ['tests/*']

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'def __repr__',
    'if settings.DEBUG',
    'raise AssertionError',
    'raise NotImplementedError',
    'if 0:',
    'if __name__ == .__main__.:',
]

[tool.pytest.ini_options]
minversion = "6.0"

[tool.poetry]
name = "intellectuals-tree"
version = "0.1.0"
description = "A web-based tool to visualize and explore the network of intellectual influences through interactive graph representation."
authors = ["Camila Morais <clmorais91@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sparqlwrapper = "^2.0.0"
black = "^24.8.0"
isort = "^5.13.2"
pylint = "^3.3.1"
ruff = "^0.6.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
